Separated out form rows into their own components for readability, merged AddSupplierModal and AddSupplierForm into one component
